+++ Jenkinsfile
@@ -179,8 +179,8 @@ node {
     	    """
 
     	    sh """
-    	    cd k8s/liara
-    	    docker build $dockerCacheArg --build-arg pip_extra_index_params="--extra-index-url ${params.PyPiRepository}" --build-arg pip_legion_version_string="==${Globals.baseVersion}+${Globals.localVersion}" --build-arg source_image="legion/base-python-image" -t legion/k8s-liara .
+    	    cd k8s/console
+    	    docker build $dockerCacheArg --build-arg pip_extra_index_params="--extra-index-url ${params.PyPiRepository}" --build-arg pip_legion_version_string="==${Globals.baseVersion}+${Globals.localVersion}" --build-arg source_image="legion/base-python-image" -t legion/k8s-console .
     	    """
 
             sh """
@@ -190,7 +190,7 @@ node {
         }
         stage('Publish Docker images'){
             if (params.PushDockerImages){
-                def images = ["legion/base-python-image", "legion/jupyterhub", "legion/k8s-edge", "legion/k8s-jenkins", "legion/k8s-grafana", "legion/k8s-edi", "legion/k8s-liara", "legion/k8s-airflow"]
+                def images = ["legion/base-python-image", "legion/jupyterhub", "legion/k8s-edge", "legion/k8s-jenkins", "legion/k8s-grafana", "legion/k8s-edi", "legion/k8s-console", "legion/k8s-airflow"]
 
                 images.each {
                     sh """
