+++ legion/legion/console/plugins/enclave_grafana/templates/enclave-grafana.html
@@ -1,0 +1,4 @@
+{% extends "layout.html" %}
+{% block content %}
+<iframe src="{{ grafana_url }}/?theme=light" style="width:100%; height:100%; border:none;"></iframe>
+{% endblock %}
