+++ legion/legion/console/plugins/homepage/templates/homepage.html
@@ -1,0 +1,5 @@
+{% extends "layout.html" %}
+{% block content %}
+<link rel="stylesheet" href="{{ url_for('homepage.static', filename='css/homepage.css') }}">
+<iframe id="dashboard" src="{{ dashboard_url }}"></iframe>
+{% endblock %}
