--- legion/legion/liara/plugins/__init__.py
@@ -1,30 +1,0 @@
-#
-#    Copyright 2017 EPAM Systems
-#
-#    Licensed under the Apache License, Version 2.0 (the "License");
-#    you may not use this file except in compliance with the License.
-#    You may obtain a copy of the License at
-#
-#        http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS,
-#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#    See the License for the specific language governing permissions and
-#    limitations under the License.
-#
-"""
-liara plugins package
-"""
-from legion.liara.plugins.homepage.view import HOMEPAGE_PLUGIN
-from legion.liara.plugins.hello.view import HELLO_PLUGIN
-from legion.liara.plugins.enclave_grafana.view import ENCLAVE_GRAFANA_PLUGIN
-
-
-def get_plugins():
-    """
-    Return array of plugins
-
-    :return list[Plugin] -- List of plugins
-    """
-    return [HOMEPAGE_PLUGIN, HELLO_PLUGIN, ENCLAVE_GRAFANA_PLUGIN]
