--- legion/legion/liara/plugins/homepage/static/css/homepage.css
@@ -1,5 +1,0 @@
-#dashboard {
-    width:100%;
-    height:100%;
-    border:none;
-}
