+++ legion/setup.py
@@ -55,7 +55,7 @@ setup(name='legion',
       license='Apache v2',
       packages=['legion'],
       include_package_data=True,
-      scripts=['bin/legionctl', 'bin/edi', 'bin/liara', 'bin/ltemplate'],
+      scripts=['bin/legionctl', 'bin/edi', 'bin/console', 'bin/ltemplate'],
       install_requires=extract_requirements(os.path.join(PACKAGE_ROOT_PATH, 'requirements', 'base.txt')),
       test_suite='nose.collector',
       tests_require=extract_requirements(os.path.join(PACKAGE_ROOT_PATH, 'requirements', 'test.txt')),
