+++ tests/resources/variables.robot
@@ -1,3 +1,3 @@
 *** Variables ***
-@{SUBDOMAINS}                       jenkins  nexus  liara
+@{SUBDOMAINS}                       jenkins  nexus  console
 @{ENCLAVE_SUBDOMAINS}               edi  edge  airflow  flower
